Moved catalog components to channelList folder
