+++ contentcuration/contentcuration/frontend/channelList/router.js
@@ -5,7 +5,7 @@ import ChannelModal from './views/Channel/ChannelModal';
 import ChannelDetailsModal from './views/Channel/ChannelDetailsModal';
 import ChannelSetList from './views/ChannelSet/ChannelSetList';
 import ChannelSetModal from './views/ChannelSet/ChannelSetModal';
-import CatalogList from './views/Catalog/CatalogList';
+import CatalogList from './views/Channel/CatalogList';
 import { RouterNames, ListTypes } from './constants';
 
 const router = new VueRouter({
